Correct timestamp comparisons
Joda DateTime objects equals() compares both the millisecond count since
the start of the epoch, and that chronologies are the the same. This is
means that two objects that represent the same instant can still be
regarded as non-equal, which is not what we want. We just want to
compare the millisecond count since the epoch.