Splitter adapter initial release
The splitter adapter manages a configurable number of other cloud
adapters. It uses the cloud adapter API to query and update the other
cloud adapters, and uses their collected responses as its own. Backend
cloud adapters can be configured with a priority which determines how
large the percentage of the pool should be allocated by each adapter,
making it possible to e.g. configure that 90% of the pool should be made
up of AWS spot instances, and 10% of regular on-demand instances.