More well-behaved AWS requests
Due to eventual consistency, we can sometimes fail to list all members
of an auto scaling group correctly: identifiers may be part of the set
returned when asking the group for its members, but those identifiers
can later not be found when a DescribeInstances call is made. This
strangeness is now handled correctly, and several other AWS requests are
now made with an exponential backoff logic, rather than a fixed delay
(as the AWS documentation dictates).