BaseCloudPool no longer supports delayed termination (instanceHourMargin).
The reason for dropping the support is four-fold: the solution (1) no longer saves any real money on most clouds, (2) has marginal (if any) real cost-saving potential, (3) unnecessarliy complicates the code, and (4) isn't flexible enough to be really useful (it assumes hour-billing).

An additional breaking change is that the victimSelectionPolicy
`CLOSEST_TO_INSTANCE_HOUR` was removed and `OLDEST_INSTANCE` and
`NEWEST_INSTANCE` were renamed to `OLDEST` and `NEWEST`, respectively
to reduce the use of EC2-jargon.